requirements.txt
altgraph==0.17.4
anyio==4.10.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
fernet==1.0.1
firebase_admin==7.1.0
Flask==3.1.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
MouseInfo==0.1.3
msgpack==1.1.1
packaging==25.0
pefile==2023.2.7
proto-plus==1.26.1
protobuf==6.31.1
pyaes==1.6.1
pyarmor==9.1.7
pyarmor.cli.core==7.6.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
PyJWT==2.10.1
PyMsgBox==1.0.9
pynput==1.8.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.1.1
pytweening==1.2.0
pywin32-ctypes==0.2.3
requests==2.32.4
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
urllib3==2.5.0
Werkzeug==3.1.3
